Split PPE methods
