+++ pfio/v2/local.py
@@ -4,29 +4,7 @@ import pathlib
 import shutil
 from typing import Optional
 
-try:
-    from pytorch_pfn_extras.profiler import record, record_iterable
-
-except ImportError:
-
-    class _DummyRecord:
-        def __init__(self):
-            pass
-
-        def __enter__(self):
-            pass
-
-        def __exit__(self, exc_type, exc_value, traceback):
-            pass
-
-    # IF PPE is not available, wrap with noop
-    def record(tag, trace, *args):  # type: ignore # NOQA
-        return _DummyRecord()
-
-    def record_iterable(tag, iter, trace, *args):   # type: ignore # NOQA
-        yield from iter
-
-
+from ._profiler import record, record_iterable
 from .fs import FS, FileStat, format_repr
